+++ src/main/java/pl/schodowski/weatherRealityCheck/model/accuWeather/Speed.java
@@ -10,7 +10,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class Speed{
     @JsonProperty("Value")
-    public double value;
+    public float value;
     @JsonProperty("Unit")
     public String unit;
     @JsonProperty("UnitType")
